requirements.txt
nano-python
nanolib
rich
# textual